refactor(ai): merge agent-call + structured into generate.ts; trim public API
streamAgentCall becomes streamText — the streaming counterpart to generateStructured, deliberately sharing the SDK facade name; internals rename to prepareGeneration / buildCallSettings so the resolution layer stops echoing transport's runProviderCall/streamProviderCall (unchanged, now private). Barrel drops the dead exports (createProviderModel, classifyProviderError, the transport calls, ProviderInstanceWithStub); getModel + callWithRetry/CallRetryError/ProviderTimeoutError stay public for pipeline fault-injection tests and the CallRetryError mapping.
